// Front/src/api/generated/model/laboratoryWork.ts
/**
 * Generated by orval v8.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { Task } from './task';

export interface LaboratoryWork {
  id?: number;
  number?: number;
  topic?: string;
  description?: string;
  maxGrade?: number;
  tasks?: Task[];
}