// Front/src/api/generated/model/task.ts
/**
 * Generated by orval v8.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { LaboratoryWork } from './laboratoryWork';

export interface Task {
  id?: number;
  number?: number;
  description?: string;
  referenceRepoPath?: string;
  lab?: LaboratoryWork;
}