// Front/src/api/generated/model/laboratoryWorkDto.ts
/**
 * Generated by orval v8.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { TaskDto } from './taskDto';

export interface LaboratoryWorkDto {
  id?: number;
  number?: number;
  topic?: string;
  description?: string;
  maxGrade?: number;
  tasks?: TaskDto[];
}