// Front/src/api/generated/model/studentSubmission.ts
/**
 * Generated by orval v8.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { Task } from './task';
import type { User } from './user';

export interface StudentSubmission {
  id?: number;
  task?: Task;
  student?: User;
  studentRepoPath?: string;
  grade?: number;
  feedback?: string;
  submittedAt?: string;
}